+++ src/clindex/indexer.clj
@@ -20,7 +20,8 @@
             [clojure.walk :as walk]
             [clojure.spec.alpha :as s]
             [clindex.specs]
-            [clindex.scanner :as scanner]))
+            [clindex.scanner :as scanner]
+            [clindex.resolve-utils :as resolve-utils]))
 
 (defn- project-facts [{:keys [:project/name :project/dependencies :mvn/version] :as proj}]
   (let [proj-id (utils/project-id name)]
@@ -65,127 +66,10 @@
                 ns-doc (into [[:db/add ns-id :namespace/docstring ns-doc]]))]
     facts))
 
-(defn- expand-symbol-alias [aliases-map current-ns-symb symb]
-  (if-let [ns-symb (namespace symb)]
-    (if-let [full-ns-symb (get aliases-map (symbol ns-symb))]
-      (symbol (name full-ns-symb) (name symb))
-      symb)
-    symb))
 
-(defn- resolve-symbol [all-ns-map ns-symb fsymb]
-  (let [ns-requires (:namespace/dependencies (get all-ns-map ns-symb))]
-    (some (fn [rns-symb]
-            (let [{:keys [:namespace/public-vars :namespace/macros] :as rns} (get all-ns-map rns-symb)]
-              (when (contains? (into public-vars macros) fsymb)
-                (symbol (name (:namespace/name rns)) (name fsymb)))))
-      ns-requires)))
-
-(defn- function-call? [all-ns-map ns-symb fq-fsymb]
-  (let [fns-symb (when-let [n (namespace fq-fsymb)]
-                   (symbol n))
-        fsymb (symbol (name fq-fsymb))]
-    (if (= fns-symb ns-symb)
-      ;; it is referring to current namespace so lets check our fns
-      (let [ns (get all-ns-map ns-symb)]
-        (contains? (into (:namespace/public-vars ns)
-                         (:namespace/private-vars ns))
-                   fsymb))
-
-      ;; it is calling other ns maybe
-      (let [ns (get all-ns-map fns-symb)]
-        (contains? (:namespace/public-vars ns) fsymb)))))
-
-(defn- macro-call? [all-ns-map ns-symb fq-fsymb]
-  (let [fns-symb (when-let [n (namespace fq-fsymb)]
-                   (symbol n))
-        fsymb (symbol (name fq-fsymb))]
-    (if (= fns-symb ns-symb)
-      ;; it is referring to current namespace so lets check our fns
-      (let [ns (get all-ns-map ns-symb)]
-        (contains? (:namespace/macros ns) fsymb))
-
-      ;; it is calling other ns maybe
-      (let [ns (get all-ns-map fns-symb)]
-        (contains? (:namespace/macros ns) fsymb)))))
-
-(defn- fully-qualify-symb [all-ns-map ns-symb symb]
-  (let [ns (get all-ns-map ns-symb)
-        ns-deps (:namespace/dependencies ns)
-        ns-alias-map (:namespace/alias-map ns)
-        ns-vars (-> #{}
-                    (into (:namespace/public-vars ns))
-                    (into (:namespace/private-vars ns))
-                    (into (:namespace/macros ns)))
-        symb-ns (when-let [s (namespace symb)]
-                  (symbol s))
-        fqs (cond
-
-              ;; check OR
-              (or (and symb-ns (contains? all-ns-map symb-ns)) ;; it is already fully qualified
-                  (special-symbol? symb)                       ;; it is a special symbol
-                  (str/starts-with? (name symb) "."))          ;; it is field access or method
-              symb
-
-              ;; check if it is in our namespace
-              (contains? ns-vars symb)
-              (symbol (name ns-symb) (name symb))
-
-              ;; check if it is a namespaces symbol and can be expanded from aliases map
-              (and (namespace symb)
-                   (contains? ns-alias-map symb-ns))
-              (expand-symbol-alias ns-alias-map symb-ns symb)
-
-              ;; try to search in all required namespaces for a :refer-all
-              :else
-              (resolve-symbol all-ns-map ns-symb symb))]
-
-    ;; transfer symbol meta
-    (when fqs (with-meta fqs (meta symb)))))
-
-(defn- fully-qualify-form-first-symb [all-ns-map ns-symb form]
-  (if (symbol? (first form))
-    (let [ns (get all-ns-map ns-symb)
-          ns-alias-map (:namespace/alias-map ns)
-          ns-vars (-> (:namespace/public-vars ns)
-                      (into (:namespace/private-vars ns))
-                      (into (:namespace/macros ns)))
-          fsymb (first form)
-          fq-symb (fully-qualify-symb all-ns-map ns-symb fsymb)]
-      (if fq-symb
-        (with-meta
-          (conj (rest form) fq-symb)
-          (cond-> (meta form)
-            (function-call? all-ns-map ns-symb fq-symb) (assoc :fn-call?    true)
-            (macro-call? all-ns-map ns-symb fq-symb)    (assoc :macro-call? true)))
-        (do
-          #_(println "[Warning] couldn't fully qualify symbol for " {:fsymb fsymb
-                                                                     :ns-symb ns-symb})
-          ;; if we couldn't resolve the symbol lets leave it as it is
-          ;; stuff like defprotoco, defrecord or protocol forms will be here
-          form)))
-    form))
-
-(defn- all-vars
-  "Return all the vars defined in all namespaces like [var-namespace var-name]"
-  [all-ns-map]
-  (->> (vals all-ns-map)
-       (mapcat (fn [ns]
-                 (let [ns-vars (-> (:namespace/public-vars ns)
-                                   (into (:namespace/private-vars ns))
-                                   (into (:namespace/macros ns)))]
-                   (map (fn [v]
-                          [(:namespace/name ns) v])
-                        ns-vars))))
-       (into #{})))
-
-(defn- split-symb-namespace [fq-symb]
-  (when fq-symb
-    (->> ((juxt namespace name) fq-symb)
-         (mapv #(when % (symbol %)))
-         (into []))))
 
 (defn- deep-form-facts [all-ns-map ns-symb form]
-  (let [is-var (all-vars all-ns-map)]
+  (let [is-var (resolve-utils/all-vars all-ns-map)]
     (loop [zloc (utils/code-zipper form)
            facts []
            ctx {:namespace/name ns-symb}]
@@ -197,7 +81,7 @@
             ;; we are deep looking at a form
             ;; lets collect this form facts
             (list? token)
-            (let [form' (fully-qualify-form-first-symb all-ns-map ns-symb token)
+            (let [form' (resolve-utils/fully-qualify-form-first-symb all-ns-map ns-symb token)
                   {ffacts :facts fctx :ctx} (form-facts all-ns-map ctx form')]
               (recur (utils/move-zipper-to-next zloc #(or (list? %) (symbol? %)))
                      (into facts ffacts)
@@ -206,8 +90,8 @@
             ;; we are deep looking at a symbol
             (symbol? token)
             (recur (utils/move-zipper-to-next zloc #(or (list? %) (symbol? %)))
-                   (let [fq-symb (fully-qualify-symb all-ns-map ns-symb token)
-                         var (split-symb-namespace fq-symb)]
+                   (let [fq-symb (resolve-utils/fully-qualify-symb all-ns-map ns-symb token)
+                         var (resolve-utils/split-symb-namespace fq-symb)]
                      (if (and (is-var var)
                               (not (:in-protocol ctx)))
                        (let [[var-ns var-symb] var
@@ -231,7 +115,7 @@
   (let [form-list' (walk/postwalk
                     (fn [x]
                       (if (symbol? x)
-                        (if-let [fqs (fully-qualify-symb all-ns-map ns-symb x)]
+                        (if-let [fqs (resolve-utils/fully-qualify-symb all-ns-map ns-symb x)]
                           (let [ns (namespace fqs)]
                             (vary-meta x merge (when (and ns x)
                                                  {:var/id (utils/var-id (symbol ns) (symbol (name x)))})))
@@ -311,7 +195,7 @@
         (d/q (+ a 4)))))
 
 
-  (fully-qualify-form-first-symb
+  (resolve-utils/fully-qualify-form-first-symb
    all-ns-map
    'clojure.data.xml.jvm.emit
    '(ns clojure.data.xml.jvm.emit
@@ -320,7 +204,7 @@
       (:require (clojure.data.xml [name :refer [qname-uri qname-local separate-xmlns gen-prefix *gen-prefix-counter*]] [pu-map :as pu] [protocols :refer [AsXmlString xml-str]] [impl :refer [extend-protocol-fns b64-encode compile-if]] event) [clojure.string :as str])
       (:import (java.io OutputStreamWriter Writer StringWriter) (java.nio.charset Charset) (java.util.logging Logger Level) (javax.xml.namespace NamespaceContext QName) (javax.xml.stream XMLStreamWriter XMLOutputFactory) (javax.xml.transform OutputKeys Transformer TransformerFactory) (clojure.data.xml.event StartElementEvent EmptyElementEvent EndElementEvent CharsEvent CDataEvent CommentEvent QNameEvent) (clojure.lang BigInt) (java.net URI URL) (java.util Date) (java.text DateFormat SimpleDateFormat))))
 
-  (fully-qualify-form-first-symb
+  (resolve-utils/fully-qualify-form-first-symb
    all-ns-map
    'clojure.data.xml.jvm.emit
    '(clojure.data.xml [name :refer [qname-uri qname-local separate-xmlns gen-prefix *gen-prefix-counter*]] [pu-map :as pu] [protocols :refer [AsXmlString xml-str]] [impl :refer [extend-protocol-fns b64-encode compile-if]] event))